tree: fold signed / and % as signed (cast both operands)
The T_SLASH/T_PERCENT constant folds cast only the left operand to signed long;
r->value stays unsigned long, so the usual arithmetic conversions make the fold
UNSIGNED and any negative operand folds wrong (the < <= > >= folds just below
already cast both):

  -7/2 -> -4 (not -3) ; 7/-2 -> 0 ; -8/-2 -> 0 ; -100%7 -> 0

Not target-specific — plain host gcc diverges on the same expression:

  unsigned long v = (unsigned long)-7, r = 2;
  (short)((signed long)v / r)               == -4
  (short)((signed long)v / (signed long)r)  == -3

+++ tree.c
@@ -713,7 +713,7 @@ struct node *constify(register struct node *n)
 				} else if (l->type & UNSIGNED)
 					value /= r->value;
 				else
-					value = (signed long)value / r->value;
+					value = (signed long)value / (signed long)r->value;
 				break;
 			case T_PERCENT:
 				if (r->value == 0) {
@@ -722,7 +722,7 @@ struct node *constify(register struct node *n)
 				} else if (l->type & UNSIGNED)
 					value %= r->value;
 				else
-					value = (signed long)value % r->value;
+					value = (signed long)value % (signed long)r->value;
 				break;
 			case T_ANDAND:
 				value = value && r->value;
